requirements.txt
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.10
python-dotenv==1.1.0
redis==6.0.0
requests==2.32.3
sniffio==1.3.1
spotipy==2.25.1
tqdm==4.67.1
urllib3==2.4.0
youtube-search-python==1.6.6
yt-dlp==2025.4.30